register_page
Готова сторінка реєстрації
Готова сторінка 404
Додано кнопки + змінила картинку для таблет та додала картинку фон для сторінки 404
