+++ src/components/AuthForm/auth-form.module.scss
@@ -1,0 +1,60 @@
+@import '../../helpers/variables.module.scss';
+@import '../../helpers/buttons.module.scss';
+
+.form__input {
+  display: block;
+  width: 100%;
+  box-sizing: border-box;
+  background: $background;
+  border: 1px solid rgba(245, 146, 86, 0.5);
+  border-radius: 40px;
+  font-style: normal;
+  font-weight: 400;
+  font-size: 14px;
+  line-height: 19px;
+  letter-spacing: 0.04em;
+  color: rgba(17, 17, 17, 0.6);
+
+  padding: 11px 14px 12px;
+  margin-top: 16px;
+  &:first-child {
+    margin-top: 0px;
+  }
+  &:nth-child(3) {
+    margin-bottom: 40px;
+  }
+}
+
+.form__button {
+  width: 100%;
+}
+
+.form__back_button {
+  display: block;
+  background-color: transparent;
+
+  margin-bottom: 5px;
+}
+
+.form__description {
+  font-style: normal;
+  font-weight: 400;
+  font-size: 12px;
+  line-height: 16px;
+  letter-spacing: 0.04em;
+  color: rgba(17, 17, 17, 0.6);
+
+  margin-top: 40px;
+}
+.description__nav {
+  color: #3091eb;
+}
+
+
+@media screen and (min-width: 768px) {
+  .form__input {
+    padding: 14px 32px 13px;
+    font-size: 18px;
+    line-height: 25px;
+  }
+}
