+++ src/components/SharedLayout/shared-layout.module.scss
@@ -1,5 +1,5 @@
 div {
-    background-color: red;
+    //background-color: red;
     width: 1000;
     height: 1000;
 }
