+++ src/pages/NotFoundPage/NotFoundPage.jsx
@@ -1,9 +1,14 @@
 import scss from "./not-found-page.module.scss";
+import { NavLink } from "react-router-dom";
 
 const NotFoundPage = () => {
 
     return (
-        <div>NotFoundPage</div>
+        <section className={scss.notfound__section}>
+            <div className={scss.container}>
+                <NavLink className={`${scss.button__primary} ${scss.notfound__button}`} to="/">&#127968; HOME</NavLink>
+            </div>
+        </section>
     )
 }
